Added Information and Hero sections
I have added the hero section along with the information section for the odin project. The images are really all just placeholders and I don't really feel as if it's useful to add others in this learning process. They would all be sized the same.

On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	new file:   dumpsterfire.png
	modified:   index.html
	new file:   jbplaceholder.png
	modified:   stylesheet.css